+++ rust/processor/src/db/postgres/migrations/2024-11-26-050428_add_market_data_to_price_feed/down.sql
@@ -3,7 +3,7 @@ DROP VIEW price_feed;
 ALTER INDEX price_feed_idx RENAME TO price_feed;
 
 CREATE OR REPLACE FUNCTION price_feed() RETURNS TABLE(
-  market_id BIGINT,
+  market_id NUMERIC,
   symbol_bytes BYTEA,
   symbol_emojis TEXT[],
   market_address VARCHAR(66),
